fix: dashboard X session
Dashboard now works. More, the layout of the head is based on the height
of the terminal!

The management of the session loading or the dashboard display at
startup is now managed in an autocommands at `VimEnter`.

lualine and clarity are disabled for dashboard. ftplugin file for
dashboard is set for prettry dashboard